Week01/pyproject.toml
[project]
name = "week01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.3.0",
    "huggingface-hub>=1.3.2",
    "openai>=2.15.0",
    "pypdf>=6.6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
